create seperate component
