+++ ui/src/App.tsx
@@ -1,6 +1,7 @@
 import PromptInput from "./components/Prompt/PromptInput";
 import PromptOutput from "./components/PromptOutput";
 import Title from "./components/Title";
+import PromptOutputTool from "./components/PromptOutputTool";
 
 function App() {
     return (
@@ -8,6 +9,7 @@ function App() {
             <Title />
             <PromptInput />
             <PromptOutput />
+            <PromptOutputTool />
         </div>
     );
 }
